docs/Release_Board-Maintainers.md: Add # column + instructions
This is for easier table maintenance, and differentiation between
primary and secondary maintainers.

I finally realized we will just need to add a '#' column if we want to
be able to easily sort this table.

I also added some (hopefully simple) instructions.  As I was not sure
of anywhere else better to put them.

Finally, I was able to easily sort (by lines) the new table,
incorporating some additions at the bottom which previously had been
out of order.